# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Bakers hub OpenAPI spec
  version: 0.0.1
servers:
- url: http://localhost:8080/v1
  description: Dev server
paths:
...
  /customers:
    get:
      tags:
      - Customers
      summary: Returns customers
      operationId: get_customers
      parameters:
      - name: status
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - active
          - not active
          - banned
        example: active
      - in: query
        name: rating
        required: false
        schema:
          type: number
        example: 7.3
      responses:
        "200":
          description: Successful response with list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.customers_controller
    post:
      tags:
      - Customers
      summary: Creates new customer
      operationId: create_customer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      responses:
        "200":
          description: Successful response with new customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.customers_controller
  /customers/{customer_id}:
    get:
      tags:
      - Customers
      summary: Returns customer with requested Id
      operationId: get_customer_by_id
      parameters:
      - name: customer_id
        in: path
        description: Customer's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      responses:
        "200":
          description: Successful response with relevant customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.customers_controller
    put:
      tags:
      - Customers
      summary: Updates existing customer
      operationId: update_customer_by_id
      parameters:
      - name: customer_id
        in: path
        description: Customer's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      responses:
        "200":
          description: Successful response with updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.customers_controller
  /customers/{customer_id}/reviews:
    get:
      tags:
      - Customers
      summary: Returns customer reviews
      operationId: get_customer_reviews_by_customer_id
      parameters:
      - name: customer_id
        in: path
        description: Customer's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      responses:
        "200":
          description: Successful response with customer review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerReview'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.customers_controller
    post:
      tags:
      - Customers
      summary: Creates customer review
      operationId: create_customer_review
      parameters:
      - name: customer_id
        in: path
        description: Customer's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerReview'
        required: true
      responses:
        "200":
          description: Successful response with list of customer reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerReview'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.customers_controller
    delete:
      tags:
      - Customers
      summary: Deletes customer review
      operationId: delete_customer_review_by_id
      parameters:
      - name: customer_id
        in: path
        description: Customer's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      - name: review_id
        in: query
        description: Review's ID
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: "1"
      responses:
        "200":
          description: Successfully deleted
          content:
            application/json: {}
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.customers_controller
components:
  schemas:
...
    Customer:
      required:
      - contact
      - name
      type: object
      properties:
        customer_id:
          type: string
          example: "1"
        name:
          type: string
          example: Igor Sergeevich
        contact:
          type: string
          example: "79201111111"
        rating:
          type: number
          example: 8.2
        status:
          type: string
          enum:
          - active
          - not active
          - banned
      example:
        contact: "79201111111"
        name: Igor Sergeevich
        rating: 8.2
        customer_id: "1"
        status: active
...
    CustomerReview:
      required:
      - author_id
      - customer_id
      - description
      - rating
      type: object
      properties:
        review_id:
          type: string
          example: "1"
        customer_id:
          type: string
          example: "1"
        author_id:
          type: string
          example: "1"
        rating:
          type: integer
          example: 9
          enum:
          - 1
          - 2
          - 3
          - 4
          - 5
          - 6
          - 7
          - 8
          - 9
          - 10
        description:
          type: string
          example: very good very well
      example:
        review_id: "1"
        rating: 9
        description: very good very well
        customer_id: "1"
        author_id: "1"
...
    DefaultError:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: integer
        message:
          type: string